+++ tests/Api/JobsTest.php
@@ -258,11 +258,41 @@ class JobsTest extends TestCase
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('post')
-            ->with('projects/1/jobs/3/play')
+            ->with('projects/1/jobs/3/play', [])
             ->willReturn($expectedArray)
         ;
 
         $this->assertEquals($expectedArray, $api->play(1, 3));
+    }
+
+    #[Test]
+    public function shouldPlayWithParameters(): void
+    {
+        $expectedArray = ['id' => 3, 'name' => 'A job'];
+        $parameters = [
+            'job_inputs' => [
+                'environment' => 'staging',
+            ],
+            'job_variables_attributes' => [
+                [
+                    'key' => 'TEST_VAR_1',
+                    'value' => 'test1',
+                ],
+                [
+                    'key' => 'TEST_VAR_2',
+                    'value' => 'test2',
+                ],
+            ],
+        ];
+
+        $api = $this->getApiMock();
+        $api->expects($this->once())
+            ->method('post')
+            ->with('projects/1/jobs/3/play', $parameters)
+            ->willReturn($expectedArray)
+        ;
+
+        $this->assertEquals($expectedArray, $api->play(1, 3, $parameters));
     }
 
     protected function getApiClass(): string